refactor: move builder to its own module
