+++ iroh/src/node.rs
@@ -7,541 +7,40 @@
 //! To shut down the node, call [`Node::shutdown`].
 use std::future::Future;
 use std::net::SocketAddr;
-use std::path::PathBuf;
 use std::pin::Pin;
 use std::sync::Arc;
 use std::task::Poll;
-use std::time::Duration;
 
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{anyhow, Result};
 use futures::future::{BoxFuture, Shared};
-use futures::{FutureExt, StreamExt, TryFutureExt};
-use iroh_bytes::downloader::Downloader;
-use iroh_bytes::store::{GcMarkEvent, GcSweepEvent, Map, ReadableStore, Store as BaoStore};
+use futures::{FutureExt, StreamExt};
+use iroh_bytes::store::ReadableStore;
 use iroh_bytes::BlobFormat;
-use iroh_bytes::{protocol::Closed, Hash};
-use iroh_gossip::net::{Gossip, GOSSIP_ALPN};
+use iroh_bytes::Hash;
 use iroh_net::derp::DerpUrl;
-use iroh_net::magic_endpoint::get_alpn;
 use iroh_net::magicsock::LocalEndpointsStream;
 use iroh_net::util::AbortingJoinHandle;
 use iroh_net::{
-    derp::DerpMode,
     key::{PublicKey, SecretKey},
     MagicEndpoint, NodeAddr,
 };
 use iroh_sync::store::Store as DocStore;
 use quic_rpc::transport::flume::FlumeConnection;
-use quic_rpc::transport::misc::DummyServerEndpoint;
-use quic_rpc::{RpcClient, RpcServer, ServiceEndpoint};
-use serde::{Deserialize, Serialize};
+use quic_rpc::RpcClient;
 use tokio::sync::{mpsc, RwLock};
 use tokio::task::JoinError;
 use tokio_util::sync::CancellationToken;
 use tokio_util::task::LocalPoolHandle;
-use tracing::{debug, error, error_span, info, trace, warn, Instrument};
+use tracing::debug;
 
-use crate::rpc_protocol::{ProviderRequest, ProviderResponse, ProviderService};
-use crate::sync_engine::{SyncEngine, SYNC_ALPN};
+use crate::rpc_protocol::{ProviderRequest, ProviderResponse};
+use crate::sync_engine::SyncEngine;
 use crate::ticket::BlobTicket;
 
+mod builder;
 mod rpc;
 
-const MAX_CONNECTIONS: u32 = 1024;
-const MAX_STREAMS: u64 = 10;
-const HEALTH_POLL_WAIT: Duration = Duration::from_secs(1);
-
-/// Default bind address for the node.
-/// 11204 is "iroh" in leetspeak <https://simple.wikipedia.org/wiki/Leet>
-pub const DEFAULT_BIND_PORT: u16 = 11204;
-
-/// How long we wait at most for some endpoints to be discovered.
-const ENDPOINT_WAIT: Duration = Duration::from_secs(5);
-
-/// Chunk size for getting blobs over RPC
-const RPC_BLOB_GET_CHUNK_SIZE: usize = 1024 * 64;
-/// Channel cap for getting blobs over RPC
-const RPC_BLOB_GET_CHANNEL_CAP: usize = 2;
-/// Default interval between GC runs.
-const DEFAULT_GC_INTERVAL: Duration = Duration::from_secs(60 * 5);
-
-/// Policy for garbage collection.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum GcPolicy {
-    /// Garbage collection is disabled.
-    Disabled,
-    /// Garbage collection is run at the given interval.
-    Interval(Duration),
-}
-
-impl Default for GcPolicy {
-    fn default() -> Self {
-        Self::Interval(DEFAULT_GC_INTERVAL)
-    }
-}
-
-/// Builder for the [`Node`].
-///
-/// You must supply a blob store and a document store.
-///
-/// Blob store implementations are available in [`iroh_bytes::store`].
-/// Document store implementations are available in [`iroh_sync::store`].
-///
-/// Everything else is optional.
-///
-/// Finally you can create and run the node by calling [`Builder::spawn`].
-///
-/// The returned [`Node`] is awaitable to know when it finishes.  It can be terminated
-/// using [`Node::shutdown`].
-#[derive(Debug)]
-pub struct Builder<D, S = iroh_sync::store::memory::Store, E = DummyServerEndpoint>
-where
-    D: Map,
-    S: DocStore,
-    E: ServiceEndpoint<ProviderService>,
-{
-    bind_port: u16,
-    secret_key: SecretKey,
-    rpc_endpoint: E,
-    db: D,
-    keylog: bool,
-    derp_mode: DerpMode,
-    gc_policy: GcPolicy,
-    rt: Option<tokio_util::task::LocalPoolHandle>,
-    docs: S,
-    /// Path to store peer data. If `None`, peer data will not be persisted.
-    peers_data_path: Option<PathBuf>,
-}
-
-const PROTOCOLS: [&[u8]; 3] = [&iroh_bytes::protocol::ALPN, GOSSIP_ALPN, SYNC_ALPN];
-
-impl<D: Map, S: DocStore> Builder<D, S> {
-    /// Creates a new builder for [`Node`] using the given database.
-    fn with_db_and_store(db: D, docs: S) -> Self {
-        Self {
-            bind_port: DEFAULT_BIND_PORT,
-            secret_key: SecretKey::generate(),
-            db,
-            keylog: false,
-            derp_mode: DerpMode::Default,
-            rpc_endpoint: Default::default(),
-            gc_policy: GcPolicy::Disabled,
-            rt: None,
-            docs,
-            peers_data_path: None,
-        }
-    }
-}
-
-impl<D, S, E> Builder<D, S, E>
-where
-    D: BaoStore,
-    S: DocStore,
-    E: ServiceEndpoint<ProviderService>,
-{
-    /// Configure rpc endpoint, changing the type of the builder to the new endpoint type.
-    pub fn rpc_endpoint<E2: ServiceEndpoint<ProviderService>>(
-        self,
-        value: E2,
-    ) -> Builder<D, S, E2> {
-        // we can't use ..self here because the return type is different
-        Builder {
-            bind_port: self.bind_port,
-            secret_key: self.secret_key,
-            db: self.db,
-            keylog: self.keylog,
-            rpc_endpoint: value,
-            derp_mode: self.derp_mode,
-            gc_policy: self.gc_policy,
-            rt: self.rt,
-            docs: self.docs,
-            peers_data_path: self.peers_data_path,
-        }
-    }
-
-    /// Sets the garbage collection policy.
-    ///
-    /// By default garbage collection is disabled.
-    pub fn gc_policy(mut self, gc_policy: GcPolicy) -> Self {
-        self.gc_policy = gc_policy;
-        self
-    }
-
-    /// Sets the DERP servers to assist in establishing connectivity.
-    ///
-    /// DERP servers are used to discover other nodes by [`PublicKey`] and also help
-    /// establish connections between peers by being an initial relay for traffic while
-    /// assisting in holepunching to establish a direct connection between peers.
-    ///
-    /// When using [DerpMode::Custom], the provided `derp_map` must contain at least one
-    /// configured derp node.  If an invalid [`iroh_net::derp::DerpMap`]
-    /// is provided [`Self::spawn`] will result in an error.
-    pub fn derp_mode(mut self, dm: DerpMode) -> Self {
-        self.derp_mode = dm;
-        self
-    }
-
-    /// Binds the node service to a different socket.
-    ///
-    /// By default it binds to `127.0.0.1:11204`.
-    pub fn bind_port(mut self, port: u16) -> Self {
-        self.bind_port = port;
-        self
-    }
-
-    /// Uses the given [`SecretKey`] for the [`PublicKey`] instead of a newly generated one.
-    pub fn secret_key(mut self, secret_key: SecretKey) -> Self {
-        self.secret_key = secret_key;
-        self
-    }
-
-    /// Whether to log the SSL pre-master key.
-    ///
-    /// If `true` and the `SSLKEYLOGFILE` environment variable is the path to a file this
-    /// file will be used to log the SSL pre-master key.  This is useful to inspect captured
-    /// traffic.
-    pub fn keylog(mut self, keylog: bool) -> Self {
-        self.keylog = keylog;
-        self
-    }
-
-    /// Set the path where known peer data is loaded on start-up and later persisted.
-    pub fn peers_data_path(mut self, path: PathBuf) -> Self {
-        self.peers_data_path = Some(path);
-        self
-    }
-
-    /// Sets the tokio runtime to use.
-    ///
-    /// If not set, the current runtime will be picked up.
-    pub fn local_pool(mut self, rt: &LocalPoolHandle) -> Self {
-        self.rt = Some(rt.clone());
-        self
-    }
-
-    /// Spawns the [`Node`] in a tokio task.
-    ///
-    /// This will create the underlying network server and spawn a tokio task accepting
-    /// connections.  The returned [`Node`] can be used to control the task as well as
-    /// get information about it.
-    pub async fn spawn(self) -> Result<Node<D>> {
-        trace!("spawning node");
-        let lp = self
-            .rt
-            .unwrap_or_else(|| LocalPoolHandle::new(num_cpus::get()));
-        // Initialize the metrics collection.
-        //
-        // The metrics are global per process. Subsequent calls do not change the metrics
-        // collection and will return an error. We ignore this error. This means that if you'd
-        // spawn multiple Iroh nodes in the same process, the metrics would be shared between the
-        // nodes.
-        #[cfg(feature = "metrics")]
-        crate::metrics::try_init_metrics_collection().ok();
-
-        let mut transport_config = quinn::TransportConfig::default();
-        transport_config
-            .max_concurrent_bidi_streams(MAX_STREAMS.try_into()?)
-            .max_concurrent_uni_streams(0u32.into());
-
-        let endpoint = MagicEndpoint::builder()
-            .secret_key(self.secret_key.clone())
-            .alpns(PROTOCOLS.iter().map(|p| p.to_vec()).collect())
-            .keylog(self.keylog)
-            .transport_config(transport_config)
-            .concurrent_connections(MAX_CONNECTIONS)
-            .derp_mode(self.derp_mode);
-        let endpoint = match self.peers_data_path {
-            Some(path) => endpoint.peers_data_path(path),
-            None => endpoint,
-        };
-        let endpoint = endpoint.bind(self.bind_port).await?;
-        trace!("created quinn endpoint");
-
-        let (cb_sender, cb_receiver) = mpsc::channel(8);
-        let cancel_token = CancellationToken::new();
-
-        debug!("rpc listening on: {:?}", self.rpc_endpoint.local_addr());
-
-        let addr = endpoint.my_addr().await?;
-
-        // initialize the gossip protocol
-        let gossip = Gossip::from_endpoint(endpoint.clone(), Default::default(), &addr.info);
-
-        // spawn the sync engine
-        let downloader = Downloader::new(self.db.clone(), endpoint.clone(), lp.clone());
-        let ds = self.docs.clone();
-        let sync = SyncEngine::spawn(
-            endpoint.clone(),
-            gossip.clone(),
-            self.docs,
-            self.db.clone(),
-            downloader,
-        );
-
-        let callbacks = Callbacks::default();
-        let gc_task = if let GcPolicy::Interval(gc_period) = self.gc_policy {
-            tracing::info!("Starting GC task with interval {:?}", gc_period);
-            let db = self.db.clone();
-            let callbacks = callbacks.clone();
-            let task = lp.spawn_pinned(move || Self::gc_loop(db, ds, gc_period, callbacks));
-            Some(AbortingJoinHandle(task))
-        } else {
-            None
-        };
-        let (internal_rpc, controller) = quic_rpc::transport::flume::connection(1);
-        let inner = Arc::new(NodeInner {
-            db: self.db,
-            endpoint: endpoint.clone(),
-            secret_key: self.secret_key,
-            controller,
-            cancel_token,
-            callbacks: callbacks.clone(),
-            cb_sender,
-            gc_task,
-            rt: lp.clone(),
-            sync,
-        });
-        let task = {
-            let gossip = gossip.clone();
-            let handler = rpc::Handler {
-                inner: inner.clone(),
-            };
-            let me = endpoint.node_id().fmt_short();
-            let ep = endpoint.clone();
-            tokio::task::spawn(
-                async move {
-                    Self::run(
-                        ep,
-                        callbacks,
-                        cb_receiver,
-                        handler,
-                        self.rpc_endpoint,
-                        internal_rpc,
-                        gossip,
-                    )
-                    .await
-                }
-                .instrument(error_span!("node", %me)),
-            )
-        };
-        let node = Node {
-            inner,
-            task: task.map_err(Arc::new).boxed().shared(),
-        };
-
-        // spawn a task that updates the gossip endpoints.
-        // TODO: track task
-        let mut stream = endpoint.local_endpoints();
-        tokio::task::spawn(async move {
-            while let Some(eps) = stream.next().await {
-                if let Err(err) = gossip.update_endpoints(&eps) {
-                    warn!("Failed to update gossip endpoints: {err:?}");
-                }
-            }
-            warn!("failed to retrieve local endpoints");
-        });
-
-        // Wait for a single endpoint update, to make sure
-        // we found some endpoints
-        tokio::time::timeout(ENDPOINT_WAIT, endpoint.local_endpoints().next())
-            .await
-            .context("waiting for endpoint")?
-            .context("no endpoints")?;
-
-        Ok(node)
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn run(
-        server: MagicEndpoint,
-        callbacks: Callbacks,
-        mut cb_receiver: mpsc::Receiver<EventCallback>,
-        handler: rpc::Handler<D>,
-        rpc: E,
-        internal_rpc: impl ServiceEndpoint<ProviderService>,
-        gossip: Gossip,
-    ) {
-        let rpc = RpcServer::new(rpc);
-        let internal_rpc = RpcServer::new(internal_rpc);
-        if let Ok((ipv4, ipv6)) = server.local_addr() {
-            debug!(
-                "listening at: {}{}",
-                ipv4,
-                ipv6.map(|addr| format!(" and {addr}")).unwrap_or_default()
-            );
-        }
-        let cancel_token = handler.inner.cancel_token.clone();
-
-        // forward our initial endpoints to the gossip protocol
-        // it may happen the the first endpoint update callback is missed because the gossip cell
-        // is only initialized once the endpoint is fully bound
-        if let Some(local_endpoints) = server.local_endpoints().next().await {
-            debug!(me = ?server.node_id(), "gossip initial update: {local_endpoints:?}");
-            gossip.update_endpoints(&local_endpoints).ok();
-        }
-
-        loop {
-            tokio::select! {
-                biased;
-                _ = cancel_token.cancelled() => break,
-                // handle rpc requests. This will do nothing if rpc is not configured, since
-                // accept is just a pending future.
-                request = rpc.accept() => {
-                    match request {
-                        Ok((msg, chan)) => {
-                            handler.handle_rpc_request(msg, chan);
-                        }
-                        Err(e) => {
-                            info!("rpc request error: {:?}", e);
-                        }
-                    }
-                },
-                // handle internal rpc requests.
-                request = internal_rpc.accept() => {
-                    match request {
-                        Ok((msg, chan)) => {
-                            handler.handle_rpc_request(msg, chan);
-                        }
-                        Err(_) => {
-                            info!("last controller dropped, shutting down");
-                            break;
-                        }
-                    }
-                },
-                // handle incoming p2p connections
-                Some(mut connecting) = server.accept() => {
-                    let alpn = match get_alpn(&mut connecting).await {
-                        Ok(alpn) => alpn,
-                        Err(err) => {
-                            error!("invalid handshake: {:?}", err);
-                            continue;
-                        }
-                    };
-                    let gossip = gossip.clone();
-                    let inner = handler.inner.clone();
-                    let sync = handler.inner.sync.clone();
-                    tokio::task::spawn(async move {
-                        if let Err(err) = handle_connection(connecting, alpn, inner, gossip, sync).await {
-                            warn!("Handling incoming connection ended with error: {err}");
-                        }
-                    });
-                },
-                // Handle new callbacks
-                Some(cb) = cb_receiver.recv() => {
-                    callbacks.push(cb).await;
-                }
-                else => break,
-            }
-        }
-
-        // Closing the Endpoint is the equivalent of calling Connection::close on all
-        // connections: Operations will immediately fail with
-        // ConnectionError::LocallyClosed.  All streams are interrupted, this is not
-        // graceful.
-        let error_code = Closed::ProviderTerminating;
-        server
-            .close(error_code.into(), error_code.reason())
-            .await
-            .ok();
-    }
-
-    async fn gc_loop(db: D, ds: S, gc_period: Duration, callbacks: Callbacks) {
-        tracing::debug!("GC loop starting {:?}", gc_period);
-        'outer: loop {
-            // do delay before the two phases of GC
-            tokio::time::sleep(gc_period).await;
-            tracing::debug!("Starting GC");
-            callbacks
-                .send(Event::Db(iroh_bytes::store::Event::GcStarted))
-                .await;
-            db.clear_live().await;
-            let doc_hashes = match ds.content_hashes() {
-                Ok(hashes) => hashes,
-                Err(err) => {
-                    tracing::error!("Error getting doc hashes: {}", err);
-                    continue 'outer;
-                }
-            };
-            let mut doc_db_error = false;
-            let doc_hashes = doc_hashes.filter_map(|e| match e {
-                Ok(hash) => Some(hash),
-                Err(err) => {
-                    tracing::error!("Error getting doc hash: {}", err);
-                    doc_db_error = true;
-                    None
-                }
-            });
-            db.add_live(doc_hashes).await;
-            if doc_db_error {
-                tracing::error!("Error getting doc hashes, skipping GC to be safe");
-                continue 'outer;
-            }
-
-            tracing::debug!("Starting GC mark phase");
-            let mut stream = db.gc_mark(None);
-            while let Some(item) = stream.next().await {
-                match item {
-                    GcMarkEvent::CustomDebug(text) => {
-                        tracing::debug!("{}", text);
-                    }
-                    GcMarkEvent::CustomWarning(text, _) => {
-                        tracing::warn!("{}", text);
-                    }
-                    GcMarkEvent::Error(err) => {
-                        tracing::error!("Fatal error during GC mark {}", err);
-                        continue 'outer;
-                    }
-                }
-            }
-
-            tracing::debug!("Starting GC sweep phase");
-            let mut stream = db.gc_sweep();
-            while let Some(item) = stream.next().await {
-                match item {
-                    GcSweepEvent::CustomDebug(text) => {
-                        tracing::debug!("{}", text);
-                    }
-                    GcSweepEvent::CustomWarning(text, _) => {
-                        tracing::warn!("{}", text);
-                    }
-                    GcSweepEvent::Error(err) => {
-                        tracing::error!("Fatal error during GC mark {}", err);
-                        continue 'outer;
-                    }
-                }
-            }
-            callbacks
-                .send(Event::Db(iroh_bytes::store::Event::GcCompleted))
-                .await;
-        }
-    }
-}
-
-// TODO: Restructure this code to not take all these arguments.
-#[allow(clippy::too_many_arguments)]
-async fn handle_connection<D: BaoStore>(
-    connecting: quinn::Connecting,
-    alpn: String,
-    node: Arc<NodeInner<D>>,
-    gossip: Gossip,
-    sync: SyncEngine,
-) -> Result<()> {
-    match alpn.as_bytes() {
-        GOSSIP_ALPN => gossip.handle_connection(connecting.await?).await?,
-        SYNC_ALPN => sync.handle_connection(connecting).await?,
-        alpn if alpn == iroh_bytes::protocol::ALPN => {
-            iroh_bytes::provider::handle_connection(
-                connecting,
-                node.db.clone(),
-                node.callbacks.clone(),
-                node.rt.clone(),
-            )
-            .await
-        }
-        _ => bail!("ignoring connection: unsupported ALPN protocol"),
-    }
-    Ok(())
-}
+pub use builder::{Builder, GcPolicy};
 
 type EventCallback = Box<dyn Fn(Event) -> BoxFuture<'static, ()> + 'static + Sync + Send>;
